separated logging subscribers to prevent duplicate log entries
