+++ src/TreeHouse/IoBundle/Bridge/WorkerBundle/Executor/ImportPartExecutor.php
@@ -8,7 +8,6 @@ use Symfony\Component\OptionsResolver\Exception\InvalidArgumentException;
 use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use TreeHouse\IoBundle\Entity\ImportPart;
-use TreeHouse\IoBundle\EventListener\ImportLoggingSubscriber;
 use TreeHouse\IoBundle\Import\ImportFactory;
 use TreeHouse\WorkerBundle\Executor\AbstractExecutor;
 use TreeHouse\WorkerBundle\Executor\ObjectPayloadInterface;
@@ -45,8 +44,6 @@ class ImportPartExecutor extends AbstractExecutor implements ObjectPayloadInterf
         $this->doctrine = $doctrine;
         $this->importFactory = $importFactory;
         $this->logger = $logger;
-
-        $importFactory->getEventDispatcher()->addSubscriber(new ImportLoggingSubscriber($logger));
     }
 
     /**
